+++ Mods ChoGGi/Buildable AttackRover/Script.lua
@@ -1,3 +1,28 @@
+-- See LICENSE for terms
+
+-- tell people how to get my library mod (if needs be)
+local fire_once
+function OnMsg.ModsReloaded()
+	if fire_once then
+		return
+	end
+	fire_once = true
+
+	-- version to version check with
+	local min_version = 34
+	local idx = table.find(ModsLoaded,"id","ChoGGi_Library")
+
+	-- if we can't find mod or mod is less then min_version (we skip steam since it updates automatically)
+	if not idx or idx and not Platform.steam and min_version > ModsLoaded[idx].version then
+		CreateRealTimeThread(function()
+			if WaitMarsQuestion(nil,"Error",string.format([[RC AttackRover requires ChoGGi's Library (at least v%s).
+Press Ok to download it or check Mod Manager to make sure it's enabled.]],min_version)) == "ok" then
+				OpenUrl("https://steamcommunity.com/sharedfiles/filedetails/?id=1504386374")
+			end
+		end)
+	end
+end
+
 function OnMsg.ClassesGenerate()
   DefineClass.AttackRover2 = {
     __parents = {