+++ Mods ChoGGi/Fix Expedition Rocket Stuck Unloading/metadata.lua
@@ -1,0 +1,14 @@
+return PlaceObj("ModDef", {
+	"title", "Fix: Expedition Rocket Stuck Unloading v0.1",
+	"version", 1,
+	"saved", 1542456000,
+	"image", "Preview.png",
+	"id", "ChoGGi_FixExpeditionRocketStuckUnloading",
+	"steam_id", "1567028510",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[This will check on load game for certain rockets stuck on the ground unable to do jack.]],
+})